Filter browser, systems, locations, sizes, top refs by path
Previously this was never filtered, and only the totals were shown.

This is a bit of a performance regression especially for longer time
periods, for a year on arp242:

master:

    dashboard           48ms  browsers
    dashboard           69ms  systems
    dashboard          102ms  sizes
    dashboard          149ms  locations
    dashboard          228ms  alltotals
    dashboard          228ms  totals
    dashboard          364ms  toprefs
    dashboard          621ms  totalpages
    dashboard          646ms  max
    dashboard          695ms  pages

This:

    dashboard          261ms  sizes
    dashboard          329ms  systems
    dashboard          339ms  browsers
    dashboard          347ms  alltotals
    dashboard          385ms  totals
    dashboard          495ms  locations
    dashboard          583ms  max
    dashboard          583ms  toprefs
    dashboard          675ms  pages
    dashboard          785ms  totalpages

Some of that is compensated with the performance enhancements of using
path_id. It's still "fast enough", and we want to make it faster we can
partition the tables or some such.